Fix i16 overflow when parsing asteroid files with MPC > 22767
The SE1 file parser read 2-byte ipl entries as i16, causing body IDs
above 32767 (e.g. 60000 for Quaoar/MPC 50000) to wrap to negative
values. This made find_file_for_jd fail to match the file, producing
"ephemeris not available" errors despite the file being valid on disk.

Read 2-byte ipl values as u16 before widening to i32.

+++ src/sweph_file/parse.rs
@@ -295,7 +295,7 @@ pub(super) fn parse_file(
         let id = if nbytes_ipl == 4 {
             r.read_i32()?
         } else {
-            r.read_i16()? as i32
+            r.read_i16()? as u16 as i32
         };
         ipl.push(id);
     }
